update id to long type

+++ hiv-treatment-support-system/backend/src/main/java/backend/doctorprofile/controller/DoctorProfileController.java
@@ -30,17 +30,17 @@ public class DoctorProfileController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Map<String, String>> update(@PathVariable int id, @RequestBody UpdateDoctorProfileRequest request) {
+    public ResponseEntity<Map<String, String>> update(@PathVariable long id, @RequestBody UpdateDoctorProfileRequest request) {
         return ResponseEntity.ok(Map.of("message", doctorProfileService.update(id, request)));
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Map<String, String>> delete(@PathVariable int id) {
+    public ResponseEntity<Map<String, String>> delete(@PathVariable long id) {
         return ResponseEntity.ok(Map.of("message", doctorProfileService.delete(id)));
     }
    
     @GetMapping("/doctor-id/{doctorId}")
-    public ResponseEntity<DoctorProfile> get(@PathVariable int doctorId) {
+    public ResponseEntity<DoctorProfile> get(@PathVariable long doctorId) {
         return ResponseEntity.ok(doctorProfileService.get(doctorId));
     }
 }
