+++ hiv-treatment-support-system/backend/src/main/java/backend/doctorprofile/dto/CreateDoctorProfileRequest.java
@@ -11,6 +11,6 @@ public record CreateDoctorProfileRequest(
 
     String biography,
     
-    int userId
+    long userId
 ){   
 }
