+++ hiv-treatment-support-system/backend/src/main/java/backend/doctorprofile/dto/UpdateDoctorProfileRequest.java
@@ -11,6 +11,6 @@ public record UpdateDoctorProfileRequest(
 
     String biography,
 
-    int userId
+    long userId
 ) {
 }
