+++ hiv-treatment-support-system/backend/src/main/java/backend/payment/dto/CreatePaymentRequest.java
@@ -11,6 +11,6 @@ public record CreatePaymentRequest (
 
     Float amount,
     
-    int scheduleId
+    long scheduleId
 ) {
 }
