+++ hiv-treatment-support-system/backend/src/main/java/backend/payment/repository/PaymentRepository.java
@@ -9,8 +9,8 @@ import org.springframework.stereotype.Repository;
 import backend.payment.model.Payment;
 
 @Repository
-public interface PaymentRepository extends JpaRepository<Payment, Integer> {
-    Optional<Payment> findByScheduleId(int scheduleId);
+public interface PaymentRepository extends JpaRepository<Payment, Long> {
+    Optional<Payment> findByScheduleId(long scheduleId);
 
     List<Payment> findByStatus(String status);
 }
