+++ hiv-treatment-support-system/backend/src/main/java/backend/schedule/dto/UpdateCheckupScheduleRequest.java
@@ -10,6 +10,6 @@ public record UpdateCheckupScheduleRequest (
     
     String status,
 
-    int doctorId
+    long doctorId
 ) {
 }
